requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.0
contextlib2==21.6.0
cryptography==43.0.0
dj-database-url==2.2.0
Django==5.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-ninja==1.2.2
django-ninja-extra==0.21.3
django-ninja-jwt==5.3.2
djangorestframework==3.15.2
gunicorn==23.0.0
iniconfig==2.0.0
injector==0.22.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
